- utils.py --> opdb.py + utils.py - test implementation for `fetch` and `insert` - example for `proposal` table
